pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Shangjin Tang", email = "shangjin.tang@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
]

[project.scripts]
python-template = "python_template:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.4.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.12",
]
